refactor: abstract VatWorkerService and move implementation to extension
