+++ packages/extension/package.json
@@ -36,6 +36,7 @@
     "@endo/eventual-send": "^1.2.4",
     "@endo/exo": "^1.5.2",
     "@endo/patterns": "^1.4.2",
+    "@endo/promise-kit": "^1.1.4",
     "@metamask/snaps-utils": "^8.3.0",
     "@metamask/utils": "^9.1.0",
     "@ocap/kernel": "workspace:^",
@@ -51,6 +52,7 @@
     "@metamask/eslint-config": "^13.0.0",
     "@metamask/eslint-config-nodejs": "^13.0.0",
     "@metamask/eslint-config-typescript": "^13.0.0",
+    "@ocap/test-utils": "workspace:^",
     "@types/chrome": "^0.0.268",
     "@typescript-eslint/eslint-plugin": "^8.1.0",
     "@typescript-eslint/parser": "^8.1.0",
