+++ packages/extension/tsconfig.json
@@ -18,6 +18,7 @@
   "include": [
     "./src/**/*.ts",
     "./src/**/*-trusted-prelude.js",
-    "./src/dev-console.js"
+    "./src/dev-console.js",
+    "./test/**/*.ts"
   ]
 }
