+++ packages/kernel/src/index.test.ts
@@ -6,7 +6,7 @@ describe('index', () => {
   it('has the expected exports', () => {
     expect(Object.keys(indexModule)).toStrictEqual(
       expect.arrayContaining(
-        ['Kernel', 'Vat', 'VatWorkerServer', 'VatWorkerClient'].concat(
+        ['Kernel', 'Vat'].concat(
           ['Cluster', 'Kernel', 'Vat'].flatMap((value) => [
             `is${value}Command`,
             `is${value}CommandReply`,
