+++ packages/kernel/src/index.ts
@@ -1,10 +1,7 @@
 export * from './messages.js';
 export { Kernel } from './Kernel.js';
 export { Vat } from './Vat.js';
-export type { VatWorker } from './vat-worker-service.js';
-export { VatWorkerClient } from './VatWorkerClient.js';
-export { VatWorkerServer } from './VatWorkerServer.js';
 export { Supervisor } from './Supervisor.js';
 export type * from './messages.js';
 export type { StreamEnvelope, StreamEnvelopeReply } from './stream-envelope.js';
-export type { VatId } from './types.js';
+export type { VatId, VatWorkerService } from './types.js';
